feat(fee-rate): 수수료율 0.1 sat/vB 지원, 수동 입력만 가능하며 기존 수수료율 선택은 integer 그대로 유지

+++ lib/providers/node_provider/node_provider.dart
@@ -310,7 +310,7 @@ class NodeProvider extends ChangeNotifier {
     return _isolateManager.broadcast(signedTx);
   }
 
-  Future<Result<int>> getNetworkMinimumFeeRate() async {
+  Future<Result<double>> getNetworkMinimumFeeRate() async {
     return _isolateManager.getNetworkMinimumFeeRate();
   }
